[proliferate] tighten session actor ownership
