+++ anyharness/crates/anyharness-lib/src/sessions/runtime/fork.rs
@@ -2,7 +2,7 @@ use anyharness_contract::v1::SessionExecutionPhase;
 
 use crate::domains::agents::model::AgentKind;
 use crate::live::sessions::actor::command::{ForkSessionCommandError, SessionCommand};
-use crate::live::sessions::connection::types::SessionStartupStrategy;
+use crate::live::sessions::actor::state::SessionStartupStrategy;
 use crate::sessions::links::model::{
     SessionLinkRecord, SessionLinkRelation, SessionLinkWorkspaceRelation,
 };
